fix: move components into the same place
