+++ src/compounds/information-panel/src/stories.tsx
@@ -5,16 +5,13 @@ import { jsx } from 'theme-ui'
 import AllThemes from '../../../utils/all-themes'
 import { CtaCell } from '../../legacy-product-table/src'
 
-import { Eligibility } from './elements/eligibility'
-
-import InformationPanel from './'
-
-const {
+import {
+  Eligibility,
   MoreInformationBlock,
   MoreInformationList,
   MoreInformationTable,
   MoreInformationText
-} = InformationPanel
+} from './'
 
 export default {
   title: 'Compounds/InformationPanel'
